requirements.txt
google-cloud-translate==2.0.1
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.2
flask-login==0.6.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
PyPDF2==3.0.1
python-docx==1.1.0 